requirements.txt
rawpy>=0.21.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
fastapi>=0.111.0
uvicorn>=0.29.0
pywebview>=4.4.0
pyinstaller>=6.5.0
